New class added in the game_api and refactoring of UI
A new class designed specially for fullfilling the rubies hash.
UI now making use of it and looking a lot clearer.

The logic behind having protected and private methods is that
protected are one idea more abstract than the private ones.
I find it a lot easier to navigate through the logic behind.